Moving

+++ Assets/Position.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Position  {
-    public int x, y;
+    public float x, y;
 
     public Position()
     {
 
     }
-    public Position(int x,int y)
+    public Position(float x, float y)
     {
         this.x = x;
         this.y = y;
